pyproject.toml
[project]
name = "deploybot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.7",
    "fastapi>=0.136.0",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-core>=1.3.0",
    "langchain-openai>=1.1.16",
    "langgraph>=1.1.9",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "sqlparse>=0.5.5",
    "tavily-python>=0.7.23",
    "uvicorn>=0.44.0",
]
